+++ src/Ninu.Emulator/Mappers/Mapper000.cs
@@ -16,19 +16,19 @@ namespace Ninu.Emulator.Mappers
             return false;
         }
 
-        public override bool TranslateProgramRomAddress(ushort address, out ushort translatedAddress)
+        public override bool TranslateProgramRomAddress(ushort address, out int translatedAddress)
         {
             if (address >= 0x8000 && address <= 0xffff)
             {
                 if (ProgramRomBankCount == 1)
                 {
                     // Handle the mirror of the first bank to the second bank.
-                    translatedAddress = (ushort)(address & 0x3fff);
+                    translatedAddress = address & 0x3fff;
                 }
                 else
                 {
                     // Assume there are just two banks.
-                    translatedAddress = (ushort)(address & 0x7fff);
+                    translatedAddress = address & 0x7fff;
                 }
 
                 return true;
@@ -38,7 +38,7 @@ namespace Ninu.Emulator.Mappers
             return false;
         }
 
-        public override bool TranslatePatternRomAddress(ushort address, out ushort translatedAddress)
+        public override bool TranslatePatternRomAddress(ushort address, out int translatedAddress)
         {
             if (address >= 0x0000 && address <= 0x1fff)
             {
